+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v8.3.10] - 2022-10-03
+- Move auth middleware registration to cortex/auth from cortex/foundation
+- Extend the default laravel AuthenticateSession middleware and simply code
+- Sync with the latest Laravel updates
+- Move SetAuthDefaults middleware to cortex/auth from cortex/foundation
+
 ## [v8.3.9] - 2022-09-06
 - Tweak entity types list 
 - Use configurable model names instead of custom ones
@@ -527,6 +533,7 @@ This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 ## v0.0.1 - 2017-03-13
 - Tag first release
 
+[v8.3.10]: https://github.com/rinvex/cortex-auth/compare/v8.3.9...v8.3.10
 [v8.3.9]: https://github.com/rinvex/cortex-auth/compare/v8.3.8...v8.3.9
 [v8.3.8]: https://github.com/rinvex/cortex-auth/compare/v8.3.7...v8.3.8
 [v8.3.7]: https://github.com/rinvex/cortex-auth/compare/v8.3.6...v8.3.7